Rename
It's kind of silly, but at least it's unique, and it's sort of a play
on the idea of "request": "please" -> "plz".  I think it's better than
calling it http-lib, because it doesn't presume to name itself as if
it were *the* HTTP library, and it's shorter.  (Although if there
weren't already a package on MELPA named "http", I'd probably just use
that.)