Change/Fix: Downcase and intern header keys
Some HTTP servers send all-lowercase header keys, which means an alist
lookup with `equal' or `string=' fails when the case differs.  We
don't want users to have to worry about this, so for consistency, we
downcase the header name.  And while we're at it, we might as well
intern it so we can use `alist-get' without having to add "nil nil
'equal" every time.