Add/Change: Handle LF-only HTTP responses
Seems that some Cloudflare responses from the matrix.org server are
using only LF characters between the headers and body (while RFC 2616
requires CRLF).  So, rather than fail completely or return invalid
data, we try to accommodate either CRLF or LF-only, but in the latter
case we display a warning (hoping that it will help to understand this
issue and can eventually be removed).