Revert "Add/Change: Handle LF-only HTTP responses"
This reverts commit 8d2654bba7bfa11525aef50f969c489c20ac0a43.

The problem was not the line endings in the server's response, but
that I wasn't aware that curl was implicitly sending an "Expect"
header, causing the server to return an additional "HTTP 100 Continue"
response.  See <https://gms.tf/when-curl-sends-100-continue.html>.  In
the next commit, we tell curl to omit that header, so the server
doesn't do that, and we only parse one response.