WIP: Use threads for sync
This is either silly, impossible, or elegant.  Regardless, it doesn't
currently work: in the plz-get-jpeg-sync test, the call to
condition-wait in the thread hangs (and freezes Emacs completely, so
it must be killed--even SIGUSR2 doesn't unfreeze it).

I don't understand why, because condition-wait releases the mutex,
which seems like it should allow the process sentinel to be called,
which should make another thread that notifies the condvar.  But
there's probably many things I don't understand about Emacs threads,
even though I've read the manual section on them a few times now.