Change/Fix: (plz-http-response-status-line-regexp) For HTTP 1 and 2
Inexplicably, using the `rx' forms `blank' and `space' worked on my
local machine but not on the GitHub CI; replacing the use of them with
plain space characters seems to fix the problem.  (My best guess is
that it's something to do with how the Nix Emacs build is configured,
Unicode settings, locale, or something like that.)

This now appears to work for both HTTP/1.1 and HTTP/2.  If possible,
I'll extend the test suite to explicitly test both (but I already
tried that and had weird errors from libcurl, so we'll see).