Add process-filter option
This allow setting the process filter via an option to the plz
function. This could already be done for asynchronous requests by
setting the process filter on the process object returned by the plz
function.

This however does not work for synchronous requests, since the plz
function only returns (or raises an error) when the process has
already run.