Widen before parsing the process buffer
When using a process filter to stream responses sometimes the
following error is raised.

```
Debugger entered--Lisp error: (wrong-type-argument number-or-marker-p nil)
  <=(200 nil 299)
  (let* ((status val)) (<= 200 status 299))
  (if (let* ((status val)) (<= 200 status 299)) (let ((status val)) (ignore status) (ignore status) (funcall (process-get process :plz-then))) (let nil (let ((err (make-plz-error :response (plz--response)))) (let* ((val (process-get process :plz-else))) (cond ((null val) (let nil (process-put process :plz-result err))) ((functionp val) (let (...) (funcall fn err))) (t (let (...) (error "No clause matching `%S'" x80))))))))
```

This only happens sometimes and I haven't found a reliable way to
reproduce it. Whenever it happens, the code was called from the
`(plz--sentinel process "finished\n")` form. Which seems to work
around a race condition that might happen. The code then runs in the
context of a narrowed process buffer.

The call to widen seem to fix the issue. I haven't seen it since.