Fix: (plz) Allow HEAD requests :as 'file
This commit avoids duplicate headers by passing --dump-header with "-"
as its value in all requests and then passing it again with the
overriding "/dev/null" value forn HEAD requests which output to the
terminal.  For HEAD requests which output to a file with --output,
--dump-header "-" is necessary since plz--http-status expects the
process buffer to always contain the headers.

This commit also consolidates the duplicate (pcase as ...) forms.